Split model.go file into multiple files by context
