+++ api/organization_model.go
@@ -1,0 +1,9 @@
+package api
+
+type Organization struct {
+	Id           string `json:"id"`
+	Name         string `json:"name"`
+	CreatedBy    string `json:"createdBy"`
+	Role         string `json:"role"`
+	IsSelfHosted bool   `json:"isSelfHosted"`
+}
