+++ api/project_model.go
@@ -1,0 +1,13 @@
+package api
+
+type Project struct {
+	IsArchived     bool   `json:"isArchived"`
+	OrganizationId string `json:"organizationId"`
+	UpdatedAt      string `json:"updatedAt"`
+	CreatedAt      string `json:"createdAt"`
+	Id             string `json:"id"`
+	Name           string `json:"name"`
+	CreatedBy      string `json:"createdBy"`
+	Role           string `json:"role"`
+	CreatedByUser  User   `json:"createdByUser"`
+}
