+++ api/ssh_key_model.go
@@ -1,0 +1,19 @@
+package api
+
+type SshKey struct {
+	User           User   `json:"user"`
+	UserId         string `json:"userId"`
+	CreatedAt      string `json:"createdAt"`
+	UpdatedAt      string `json:"updatedAt"`
+	Id             string `json:"id"`
+	Name           string `json:"name"`
+	Description    string `json:"description"`
+	OrganizationId string `json:"organizationId"`
+	Value          string `json:"value"`
+}
+
+type SshKeyCreatePayload struct {
+	Name           string `json:"name"`
+	OrganizationId string `json:"organizationId"`
+	Value          string `json:"value"`
+}
